# image: p_median3x3: new sort algrthm,~185x speedup