The included routine is stand-alone (does not depend on external routines).
The biggest hit to performance in the previous code was the use of the memcpy
system call, which has been removed.

A clever reduction in the number of sorting comparisons and swaps was made as
well as a clever comparison with integers, reducing instruction overhead. The
overall performance is closer in line with the other image stencil routines
now.  The total program size was also reduced by approximately 1824 bytes when
calling this routine compared to the previous code. Performance benchmarked
around 105 MegaPixels/second which I feel is pretty good considering the all
the sorting and logical operations that don't require much math.

Signed-off-by: James Ross <james.a.ross@gmail.com>