@olajep 2015-08-27
Thank you James,

Where is that 185x speedup number coming from, this is what I get, what am I doing wrong?
```
p_median3x3

Epiphany
CFLAGS="-O2 -g -msoft-cmpsf -ffast-math -mfp-mode=round-nearest -ffp-contract=fast"
Before
;name, size, duration (ns)
p_median3x3_f32, 512, 226913

After
;name, size, duration (ns)
p_median3x3_f32, 512, 37933

Speedup: 6

ARM
CFLAGS="-O3 -ffast-math -mcpu=cortex-a9 -mfpu=neon"
Before
;name, size, duration (ns)
p_median3x3_f32, 16384, 3815455

after
;name, size, duration (ns)
p_median3x3_f32, 16384, 2691080

Speedup: 1.40


x86
CFLAGS="-O3 -march=native -mcpu=native -ffast-math"
before
;name, size, duration (ns)
p_median3x3_f32, 16384, 262656

after
;name, size, duration (ns)
p_median3x3_f32, 16384, 193952

Speedup: 1.35
```