@jar 2015-08-27
Thanks for the tip.  It seems that the benchmark summary table should load
memcpy and other routines for a better comparison. And in the benchmark
numbers in your previous email use -O2 for Epiphany.  I'm using -O3. While
the actually performance numbers aren't clear anymore, it is certainly a
greatly improved code no matter how you measure.
On Aug 27, 2015 11:42 AM, "Ola Jeppsson" <notifications@github.com> wrote:

> I guess you pass in -Wl,-T,/opt/adapteva/esdk/bsps/current/fast.ldf to
> LDFLAGS
> This will place some of the functions e.g., memcpy in external RAM which
> is sloow.
>
> Try the internal.ldf linker script instead
>
>
> // Ola
> On 2015-08-27 16:31, jar wrote:
> > It is my pesonal benchmark peformance difference between the previous
> code
> > running on the Epiphany and this code. If you look at the summary table,
> > the old result was an outlier and particularly slow. Any of the codes
> that
> > use memcpy are extremely slow. That change alone is responsible for most
> of
> > the performance improvement on Epiphany. The rest of the gains were from
> > removing superfluous copies, inlining, and reduction of operations.
> > -JAR
> > On Aug 27, 2015 10:15 AM, "Ola Jeppsson" <notifications@github.com>
> wrote:
> >
> > > Thank you James,
> > >
> > > Where is that 185x speedup number coming from, this is what I get,
> what am
> > > I doing wrong?
> > >
> > > p_median3x3
> > >
> > > Epiphany
> > > CFLAGS="-O2 -g -msoft-cmpsf -ffast-math -mfp-mode=round-nearest
> -ffp-contract=fast"
> > > Before
> > > ;name, size, duration (ns)
> > > p_median3x3_f32, 512, 226913
> > >
> > > After
> > > ;name, size, duration (ns)
> > > p_median3x3_f32, 512, 37933
> > >
> > > Speedup: 6
> > >
> > > ARM
> > > CFLAGS="-O3 -ffast-math -mcpu=cortex-a9 -mfpu=neon"
> > > Before
> > > ;name, size, duration (ns)
> > > p_median3x3_f32, 16384, 3815455
> > >
> > > after
> > > ;name, size, duration (ns)
> > > p_median3x3_f32, 16384, 2691080
> > >
> > > Speedup: 1.40
> > >
> > >
> > > x86
> > > CFLAGS="-O3 -march=native -mcpu=native -ffast-math"
> > > before
> > > ;name, size, duration (ns)
> > > p_median3x3_f32, 16384, 262656
> > >
> > > after
> > > ;name, size, duration (ns)
> > > p_median3x3_f32, 16384, 193952
> > >
> > > Speedup: 1.35
> > >
> > > —
> > > Reply to this email directly or view it on GitHub
> > > <https://github.com/parallella/pal/pull/217#issuecomment-135447680>.
> > >
> >
> > —
> > Reply to this email directly or view it on GitHub <
> https://github.com/parallella/pal/pull/217#issuecomment-135453221>.
> >
>
>
> --
> Phone: +46.733208642
> Skype: olajep
> Email: ola@adapteva.com
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/parallella/pal/pull/217#issuecomment-135472929>.
>